Moving Nomad processors to the right module
