+++ dynamic-config/server/configuration-provider/pom.xml
@@ -68,6 +68,11 @@
       <artifactId>packaging-support</artifactId>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>org.terracotta.internal</groupId>
+      <artifactId>management</artifactId>
+      <scope>provided</scope>
+    </dependency>
 
     <!-- Access TcServerMain from TcServer and the exception classes -->
     <dependency>
