+++ dynamic-config/server/configuration-provider/src/main/java/org/terracotta/dynamic_config/server/configuration/service/NomadServerManager.java
@@ -36,13 +36,13 @@ import org.terracotta.dynamic_config.server.configuration.nomad.ConfigChangeAppl
 import org.terracotta.dynamic_config.server.configuration.nomad.NomadServerFactory;
 import org.terracotta.dynamic_config.server.configuration.nomad.UncheckedNomadException;
 import org.terracotta.dynamic_config.server.configuration.nomad.persistence.NomadConfigurationManager;
-import org.terracotta.dynamic_config.server.configuration.nomad.processor.ApplicabilityNomadChangeProcessor;
-import org.terracotta.dynamic_config.server.configuration.nomad.processor.ClusterActivationNomadChangeProcessor;
-import org.terracotta.dynamic_config.server.configuration.nomad.processor.DefaultRoutingNomadChangeProcessor;
-import org.terracotta.dynamic_config.server.configuration.nomad.processor.MultiSettingNomadChangeProcessor;
-import org.terracotta.dynamic_config.server.configuration.nomad.processor.NodeAdditionNomadChangeProcessor;
-import org.terracotta.dynamic_config.server.configuration.nomad.processor.NodeRemovalNomadChangeProcessor;
-import org.terracotta.dynamic_config.server.configuration.nomad.processor.SettingNomadChangeProcessor;
+import org.terracotta.dynamic_config.server.configuration.service.nomad.processor.ApplicabilityNomadChangeProcessor;
+import org.terracotta.dynamic_config.server.configuration.service.nomad.processor.ClusterActivationNomadChangeProcessor;
+import org.terracotta.dynamic_config.server.configuration.service.nomad.processor.DefaultRoutingNomadChangeProcessor;
+import org.terracotta.dynamic_config.server.configuration.service.nomad.processor.MultiSettingNomadChangeProcessor;
+import org.terracotta.dynamic_config.server.configuration.service.nomad.processor.NodeAdditionNomadChangeProcessor;
+import org.terracotta.dynamic_config.server.configuration.service.nomad.processor.NodeRemovalNomadChangeProcessor;
+import org.terracotta.dynamic_config.server.configuration.service.nomad.processor.SettingNomadChangeProcessor;
 import org.terracotta.json.ObjectMapperFactory;
 import org.terracotta.nomad.server.NomadException;
 import org.terracotta.nomad.server.UpgradableNomadServer;
