+++ dynamic-config/server/configuration-repository/pom.xml
@@ -71,10 +71,5 @@
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.terracotta.internal</groupId>
-      <artifactId>management</artifactId>
-      <scope>provided</scope>
-    </dependency>
   </dependencies>
 </project>
